fix: attach activity card to app rail bell
Generated-By: PostHog Code
Task-Id: 4a1c53d9-0ca7-4ebe-9936-0c6dab92ca96
